Minor changes
- Updated tox and .yml.
- Organized into IO files.
- Renamed read_to_df
- buff.. turned private.
