+++ .github/workflows/ci.yml
@@ -9,16 +9,36 @@ jobs:
 
     runs-on: ubuntu-latest
     strategy:
-      matrix:
-        python-version: [3.7, 3.8, 3.9]
+      # continue even if an enviroment fails
+      fail-fast: false
+
+      matrix: 
+        python-version: [3.9]
+        tox_env:
+          - style
+          - docstyle
+          - check-manifest
+          - coverage
+
+        # check version by version
+        include:
+          - py_version: 3.7
+            tox_env: py3.7
+          - py_version: 3.8
+            tox_env: py3.8
+          - py_version: 3.9
+            tox_env: py3.9
 
     steps:
       - uses: actions/checkout@master
+
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
         with:
-          python-version: ${{ matrix.python-version}}
+          python-version: ${{ matrix.python-version }}
+
       - name: Install tox
-        run: pip install tox tox-gh-actions
-      - name: Test with tox
-        run: tox -r
+        run: pip install tox
+        
+      - name: tox -e ${{ matrix.tox_env }}
+        run: tox -re ${{ matrix.tox_env }}
