+++ deepspyce/__init__.py
@@ -25,4 +25,4 @@ __version__ = "0.0.1"
 # IMPORTS
 # =============================================================================
 
-from .core import *  # noqa
+from .io import *  # noqa
