+++ tox.ini
@@ -1,15 +1,9 @@
 [tox]
-envlist = python3.7, py38, py39, style, docstyle, check-manifest, coverage
+envlist = py3.7, py3.8, py3.9, style, docstyle, check-manifest, coverage
 
 # =============================================================================
 # CONF
 # =============================================================================
-
-[gh-actions]
-python =
-    3.7: python3.7
-    3.8: py38, style, docstyle, check-manifest, coverage
-    3.9: py39
 
 #[pytest]
 #addopts = -n "auto"
@@ -31,6 +25,7 @@ deps =
     flake8
     flake8-import-order
     flake8-black
+    flake8-builtins
     pep8-naming
 commands =
     flake8 setup.py tests/ deepspyce/ {posargs}
@@ -49,8 +44,8 @@ deps =
     pytest-cov
 commands =
     - coverage erase
-    - pytest -q tests/ --cov=deepspyce --cov-append --cov-report=
-    coverage report --fail-under=80 -m
+    - pytest -v tests/ --cov=deepspyce --cov-append --cov-report=
+    coverage report --fail-under=85 -m
 
 [testenv:check-manifest]
 skip_install = True
